refactor: refactored twitter scripts, removed cron stuff
